kernel: Don't load group history when in noshell mode
Loading of group_history synchornizes with kernel_sup in order
to use disk_log. Doing that when in the user process can lead to
a deadlock if logger (or anything really) tries to write to
standard out at the same time. Normally the deadlock will solve
itself when the kernel application is started, but if it should
fail to start for some reason user is terminated before it can
print anything and logger will print the log messages to standard
error instead.