[erts] Priority messages - Implementation of priority message queue
The message queue (MQ) can now begin with a priority message queue (PMQ)
where priority messages can be appended. The priority message queue is
identified using a priority message queue start marker (PMQS) and a priority
message queue end marker (PMQE). When a priority message is received it is
inserted just before the PMQE marker.

MQ -> PMQS -> PrioMsg1 -> ... -> PrioMsgN -> PMQE -> Msg1 -> ... -> MsgN

The receive expression will try to match messages from the beginning of
the MQ going through the messages in PMQ one by one continuing with
non-priority messages when it has reached the end of the PMQ.

The selective receive optimization will continue to work. When there are
any receive markers present, an incoming priority message will be
scanned for receive marker references. If not found, the receive can
continue as usual. If a reference found in the priority message corresponds
to the currently used receive marker, the priority message will be matched
against before continuing with the next ordinary message.

If a priority message contains a reference corresponding to a receive
marker and the receive marker is used in multiple receive expressions,
the second and following receive expressions will have to scan the whole
PMQ before continuing to the location of the receive marker. A PMQ receive
marker could be inserted preventing this. However, the PMQ is expected to
be small, and it is quite unexpected that a reference is part of a priority
message, so this has not been implemented.