shell: Implement lazy-read for noshell
When running in `-noshell` mode, it is "normal" for programs
to not eagerly read from stdin. This commit changes so that
we only issue read requests to stdin when `user` has received
a read request from the application.

This change also makes it so that on Windows we don't have to
abort an ongoing read request, which means that we can use
ReadFile in non-overlapped mode and thus Unicode will work
when reading from stdin in noshell mode on windows.