stdlib: add type specs to shell docs
adds type specifications to shell_docs. allows users to type `ht(ssh,
daemon)` and get the documentation together with the spec.

the types shown are those included in the `beam_lib:chunks`, which
currently shows some types that are not exported by needed by functions.
these types are the same ones shown on the online docs.

the shell_docs_SUITE.erl has been parallelised due to the amount of time
taken to render the new documentation. the reason is that we need parse
the debug_info from a beam file, where we previously were not doing this
and thus, not showing the type information.

when the emulator does not have enough SMP, we skip a specific test
case, otherwise it can take longer than 40 min to finish the testcase.
we have only observe this case taking longer than 20 min in cases where
the emulator starts with a single scheduler.