public_key: Do not hide crypto runtime exceptions
This type of runtime error hiding is not considered something
that is part of the API that needs to be backwards compatible.
This only hides the real error and does not help prevent bugs.
This exceptions occur when input data is not correct and
should be clearly reported when it occurs.