ssh: fix Opts creation in ssh_acceptor:listen
This was [{active, false}, {reuseaddr,true} | ?GET_OPT(socket_options,
Options)] in ssh_acceptor, which is wrong. For one, it would prevent
the usage of option inet_backend since this option must be the first
in the list. For another, in options given to gen_tcp or ssl, the last
option wins, such that if for example {active, true} was in the list
of given socket options, it would be set to {active, true} despite the
tacked-on {active, false}.

Co-authored-by: Maria Scott <maria-12648430@hnc-agency.org>
Co-authored-by: Jan Uhlig <juhlig@hnc-agency.org>