Clarify burst mode protection and non file based logging
Reading the docs my impression was that the reasons mentioned
for burst mode protection do only apply to file-based logging or
at least the examples only do.

Locking to the console is fairly common in cloud-based
environments.
I have deactivated the burst mode protection (that had swallowed
about 2/3s of our log messages) and the system is running fine
without any impact on CPU or memory that I could spot.

So, this is part me wanting to clarify this in the doc but also
part me checking in about my understanding of the interplay of
this.