Provide an overview of logger overload protection
Digging into the topic of logger overload protection I found it
somewhat hard to follow and keep track of what all the systems
do on a high level.

It probably doesn't solve this for everything, but I think
providing a high level overview of the measures in place can
help others to better understand the mechanisms and dive into
the one they care the most about.