Add support for `SIGINFO` system signal
This signal is emitted when a user type Ctrl+T in a terminal. This is
only supported by few operating systems such as *BSD and OSX.

By default, the operating system displays on stdout the system process
PID, the system load and the duration the process is running. The signal
allows the application to display more details about its status and
progress.