Take make(1) executable name from `$MAKE` environment variable
[Why]
A few places have the executable name hard-coded to `make`. However, the
Makefiles require GNU make, which is might be available as another name.
For instance, on *BSD where make(1) is another implementation, GNU make
is available ase `gmake`.

This was not a problem when building Erlang/OTP. But it was when running
the tests.

[How]
The make variable `$(MAKE)` which always contain the name of (or path
to) the make executable being used, is passed to the shell scripts. The
shell scripts then use this variable, defaulting to `make` if necessary.

In `lib/common_test/test_server/configure.ac`, the order in which the
make executable is detected is swapped: `gmake` is tried first, then
`make`.