Update attributes in EEP 48
Currently, editors and documentation tooling must
parse metadata from several chunks in order to
function properly. The source is in the compilation
chunk. The behaviours a module implements is in the
attributes chunk. And the annotation of any definition
requires traversing abstract code.

With this commit, we introduce three new attributes,
`:behaviours`, `:source`, and `:source_anno`, so
this additional data can be standardized and easily
located.

We also remove `edit_url` from the standard. As it is
not used by any of the official tooling, languages, or
libraries.