kernel: fix bug with composing unicode character
When the first charaacter is a composing unicode character like
ZWNJ ZWJ characters or U+0308 it binds with the last character
that is part of the prompt, and creates a grapheme cluster.
In this case we need to split up the grapheme cluster and remove
the composing character and keep the character part of the prompt.