compiler: Use strict generators in comprehensions
Since the compiler is an important application, it is worthwhile using
strict generator operators where possible to help find mistakes
quicker, and also make it clearer at a glance whether a generator with
a pattern is filtering based on the pattern or is keeping all elements.

If I have counted correctly, there are now 130 strict list and map
generators with failable patterns and 73 relaxed map and list
generators with failable patterns. There is one strict binary
generator.