sasl: include appup files in release package
This commit changes the behaviour of systools:make_tar/1,2 so that
application upgrade (.appup) files for all included applications are
included in the release package.

To support just-in-time (JIT) creation of release upgrade specifications
on a target system, using systools:make_relup/3,4, it is necessary to
have the application upgrade files (.appup) installed.

See this community forum discussion thread on Day 1 or Day 2 (JIT)
release specification creation:
https://erlangforums.com/t/release-upgrade-specifications-day-1-or-day-2/4013