compiler: Note non-obvious behavior of beam_ssa:split_blocks/4
beam_ssa:split_blocks/4 does not split and create a new empty block if
the predicate returns true for the first instruction in a block. As I
have gotten tripped up on this for the second time, the first time was
at least a year ago, note this in the function comment.