Simplify extraction of binary segments of dynamic size
The flag argument for the bs_get_get_binary_2() helper function is not
used and can be removed.

In optimized code for OTP 26 and later, the size operand for the BEAM
instruction bs_get_binary_2 is always a BEAM register. Therefore, we
can emit smaller and less performant code when the size operand is an
integer or the atom `all`.