Simplify sign extension for integer segments
Simplify sign extension in a way that allows some C compilers
(Clang at the time of writing) to emit branchless code. Other
compilers (GCC at the time of writing) will emit a single branch
instruction, but the rest of the code is simpler than the previous
approach to sign extension.