erts: Fix possible Visual Studio bug
Visual Studio 17 seems to compile the original code
into something that make coverage into MAX_SMALL
instead of coverage_data. Extracting the data into
a variable seems to fix the error, don't know why
but hopefully it will be fixed in the compiler someday.
Or maybe the original code is somesort of undefined
behaviour I'm unable to see.

The original code compiles as it should in Visual Studio 14.