erts: Include jit.pdb and release all pdb files
When debugging on windows, the name of the pdb file is
not the name of the file being debugger, but the name
of the file as it was compiled. We rename beam.jit.dll
to beam.smp.dll during the build so the pdb needed is called
jit.pdb and not smp.pdb.

We also make sure to release all pdb files so that they
are part of the installer.