ssl: Peer ext-keyusage extension should alwyas be verified if present
Critical extension is not relevant in this context, when we
know the extension it should always be verified. If present
in a itermidate CA the extension may also be acceptable
if the usage is relaxed by the special "any" value, option
allow_any_ca_purpose is added to configure that.