otp: Fix notes links to render correctly
ex_doc (or rather EarmarkParser) does not consider the
whitespace between reference links significant so joins
multiple references into a single link. Many MD standards
seem to work this way, though the majority do not.

Anyway we need to fix this so that the links render properly,
see https://github.com/RobertDober/earmark_parser/issues/159
for details on when/if this issue is solved.