Optimize checking the number of empty queues in try_steal_task
Currently we're checking it in every iteration in the fast loops (and
this shows in profiling), and not at all in the final loop (which is
the only one that can be slow enough to deserve it). This just fixes
this small point I had forgotten in the last commit.