stdlib: Fix preorder and postorder traversals in `digraph_utils`
The previous `digraph_utils:preorder/1` and `digraph_utils:postorder/1`
did not start the traversal from root nodes. This fix makes both
traversals only start or restart from a root node in one of the
components, or an arbitrary node if no root node can be visited.
Since in the previous version, the search can start and restart from
arbitrary nodes, this fix should not break backwards compatibility.

It is worth noting that the digraph does not have a notion of a left or
right subtree. The result of both traversals may not be the same after
relabelling vertices.