[erts] Priority messages - Implementation of priority marked signals
The action performed when an ordinary message is received, is to append it
to the end of the message queue. When receiving a priority marked message
that action might not be taken, but instead an alternate action might be
taken, i.e, the message might be appended to the end of the priority message
queue. There currently exist one type of alternate action messages, namely
messages sent using an alias.

This commit generalize the alias message passing functionality to passing
an alternate action signal which can be used by priority messages, priority
exit signals as well as alias messages (and potentially other future
alternate action messages).

Both local and distributed alternate action signals are supported.