Add `process_info(Pid, label)` for retrieving the process label
The only documented way to retrieve the label for a process was
by calling `proc_lib:get_label/1`.

This commit teaches `process_info/2` the option `label` for retrieving
the process label. This is especially useful when one needs to
retrieve other process info items at the same time. For example:

    process_info(Pid, [label,registered_name])