compiler destructive update: Fix bug in handling of nested patches
While refactoring the code responsible for merging patches to
different elements of a literal in function calls in order to reuse it
to solve issue #9100, it was discovered that it only correctly handle
patches to top-level elements. If multiple elements of a tuple inside
another tuple was patched, it would crash.

This commit fixes the above mentioned bug and also refactors the code
to separate the parts that are specific to handle function arguments
from the parts that merges patches to literals.

The correctness of the refactoring is tested by already existing
tests. A test case verifying the fix will be added in the patch
addressing #9100 as is it not worthwhile to spend the time to
construct a test-case testing each of the issues separately.