kernel: Update standard_error to use prim_tty nif
The primary reason that we do this is to make writing to standard_error
as synchronous as possible. When using the fd driver the write was
done on an async thread, which meant that there was no simple way
to get an acknowledgement when the write had completed. Now that we
use a NIF we can much more easily get a check when the write is done
so that we know when io:format(standard_error, ...) has returned we
have also written to that fd.