argparse: Accept the progname in the command path
[Why]
The documentation of `cmd_path()` states that it always starts with the
progname. Indeed, `argparse:parse/{1,2}` returns a command path with the
progname at the beginning.

However, if `argparse:help/2` is called with `#{command => CmdPath}` in
the parser options with `CmdPath` starting with the progname, it crashes
with a `badkey` exception because it expects the command path to only
contain commands and sub-commands.

[How]
The patch drops the first element of the command path if it matches the
progname.

This adds support for the correct values of `CmdPath` while retaining
backward compatibility with callers that dropped the progname themselves
to work around the issue.