Makefile.in: Export $MAKE to be used by other scripts
[Why]
GNU make is sometimes installed as `gmake` and `make` points to another
implementation that does not support GNU make extensions.

The name of the make command is stored in `$(MAKE)` and sub-make
processes automatically inherit and use this variable. However other
executed commands and scripts do not get this variable by default.
Therefore of a script tries to call make(1) again, it will either not
find this value or, more likely, will hard-code `make`.

[How]
The top-level Makefile now exports the value of `$(MAKE)` into the
environment of all executed sub-shells. This way, scripts can use the
correct make command name if they need to run make(1) again.