erts: Fix local_to_universal time without dst
With glibc 2.37, the behaviour of mktime was changed to
no longer return an error if tm_isdst = 1 but the TZ actually
does not have any dst data. For example if TZ=UTC.

Instead of trusting the timezone, and not assuming that dst
is in effect, glibc decided to trust the caller and assume
that dst is in effect. This means that Erlang and glibc now
have different behaviours in this regard. So in order to
keep backward compatability with how Erlang has always worked
we instead check is the TZ has daylight handling at all, and if
not we set isdst to 0.

The change was introduced by this commit: https://sourceware.org/git/?p=glibc.git;a=commit;h=83859e1115269cf56d21669361d4ddbe2687831c and
a bug report was filed with glibc here:
https://sourceware.org/bugzilla/show_bug.cgi?id=31144, but
it seems like the new behaviour is desired.