erts: Use "TZ" env var for selecting time zone
According to POSIX.1 when selecting which time zone a
system should run in, it should first look at the "TZ"
environment variable. This is what mktime and friends
do on Unix+Windows, but as Erlang has its own implementation
of mktime this was not true.

This commit reads the environment variable and from
that looks up the required information from the registry.