Fix expansion of symbolic local funs
If a 'fun f/n' used the same name and arity f/n as a BIF 'erlang:f/n',
for example 'max/2', the symbolic fun would get unnecessarily
rewritten as 'fun (X1, ..., Xn) -> f(X1, ..., Xn) end'.

In addition, the rewritten fun body was then processed as a normal
call, so if 'max/2' was declared as an import from some module m, it
got expanded to a call to 'm:max/2', making imports "work" for
symbolic funs but only if their names look like BIF names in order to
sneak past the linter.

This change makes the expander do the right thing for all imports,
auto or explicit, creating symbolic remote funs. It is up to the
linter (standard or modified) to decide whether it is allowed.