implement stack_size match spec action
This PR implements a `stack_size` match spec action. It gets the size of the process stack in the same way as `erlang:process_info(Pid, stack_size)`.

This action is useful when trying to reliably reconstruct how the process stack looks at any point in time, without having to request a full `stack_trace` in the match spec at each call.

## Why this is useful
Erlang has tail call elimination, which leads to stack frames being elided if they have another function call in the tail position. The tracing facility has options to enable both `call` and `return_to` messages, but these are not symmetrical. (there may be more `call`s vs `return_to`s in a trace when TCs are performed)

This makes it tricky/impossible to know exactly what the state of the stack looks like in a process at any point in time by just looking at a sequence of trace messages.

This can be remediated by asking for a stack trace on each trace event (the `stack_trace` match spec action), but this is expensive and unnecessary when you simply want to be able to estimate the size of the call stack at any point in execution.

The primary use case of this feature for me is being able to more reliably render flame graphs from traces.