ssh: improve disconnectfun behaviour and documentation
The log message that is passed as argument to disconnectfun arity 1
is now excluded from the arity 2 function. Instead a reason in the
format of an atom is used and textual representation of the error
always are provided in the details field of the Info map
(second argument of the callback).

This documentation of the disconnectfun is also extended.