Reimplement escript archives
The support for zip archives for packaging Erlang applications is
an experimental features, and as such can be subject to incompatible
changes.

The support for archives in the code server and the `erl_prim_loader`
module complicates the implementation and make code loading slower
even when no archives are used. Furthermore, the existence of the
feature prevents or vatstly complicates further optimization of code
loading.

Therefore, this commit removes the support for code archives from
the code server, and from the `code` and `erl_prim_loader` modules.

Having multiple BEAM files and other files in an `escript` is still
supported, but in an incompatible way for escripts created by
Erlang/OTP 28. We decided to do it like this to keep
`escript:create/2` and `escript:extract/2` as simple as possible, and
make it possible to phase out the old archive format as soon as
possible.

However, escripts created by Erlang/OTP 27 and earlier will run
in Erlang/OTP 28, provided that they don't rely on archive features
in the code server or `erl_prim_loader`.

Co-authored-by: Björn Gustavsson <bjorn@erlang.org>