Implement `select_read`
We need to differ between a complete read, e.g for a stream socket
with specified length during `socket:recv/*` that collects the
requested amount, and an incomplete read where the OS returns
partial data.  So `select_read` will mean that the read is complete
but we are awaiting a `select` message due to the `select_read`
option being active, while `select` means that the read is incomplete
and we are waiting for a `select` message.