diameter: DiameterURI transport defaults to tcp
When `rfc` is missing in the option, the transport part of the
diameter_uri becomes an empty atom (''), which causes the diameter
module to fail to re-encode the decoded AVP (because it won't match
diameter_types:'DiameterURI'/3 properly).

This sets `tcp` as the default transport as described in RFC6733
instead of leaving it empty.