Skip test pcre incompat for Recursive backreferences
I think this is explained by

PCRE:
Back references of this type cause the group that they reference
to be treated as an atomic group. Once the whole group has been matched,
a subsequent matching failure cannot cause backtracking into the middle
of the group.

PCRE2:
This restriction no longer applies, and backtracking into such groups
can occur as normal.