Limit peak memory use while loading code on small systems
In order to start the runtime system as fast as possible, modules
known to be always needed to be loaded are loaded in parallel.
When those modules have been loaded, the `error_handler` mechanism
kicks in to load modules on demand.

However, loading code in parallel will increase the peak memory usage
while loading, which can be serious issue for embedded system with a
limited amount of memory.

This commit adds an heuristic in the `init` module to disable parallel
loading when it seems probable that the system has limited
memory. Parallel loading is disabled if running on a 32-bit CPU with
a single core. A few calls to `erlang:garbage_collect/0` are also
added to make sure that binaries holding BEAM code are deallocated as
soon as possible.

While at it, the number of processes spawned by `erl_prim_loader` when
fetching and preparing BEAM files has been changed to 1.5 times the
number of schedulers, and capped at 32 processes. It used to be 32
processes regardless of the number of cores.