gh: Migrate PR test logs and docs to erlang.org
Github has a limitation of 10GB of data that you can have in
a github pages instance, and we constantly hit that limit for
our PRs. So the PR viewing service has not been migrated to
erlang.org. This commit redirects the links and removes the
infrastructure to upload PR results.