erts: Remove faulty assert checking for DELAYED_DEL_PROC flag
In Erlang/OTP 22 functionality was added to yield after process "death"
in order to send large error reasons for down/exit signals. This change
meant that we not can be in state FREE, but the DELAYED_DEL_PROC is first
set later, so the assert checking for it is incorrecy.