WIP: Forward parent death to descendant processes (unix)
If the forker's connection to the parent BEAM is broken or closed, react
by killing all spawned children.  When a spawned port is closed, kill
the associated OS process.

A concise demonstration of the problem being solved is to run the
following command with and without the patch, then kill the BEAM.
Without the patch, the "sleep" process will continue:

    erl -noshell -eval 'os:cmd("sleep 60")'

TODO:
* Needs a decision made between killing the process or process group.
* Separate patch for win32