Write failing tests for child process cleanup
These tests demonstrate that the VM allows a spawned child OS process
to outlive it.  Current behavior relies on the child to "politely"
watch stdin and exit when the file descriptor is close.  Some but not
all external processes follow this rule.

The next few patches will change VM behavior to kill all child
processes on exit, which should fix the tests introduced here.