[application_controller] Stop starting applications when terminating node
Applications that didn't finish starting could be left in a half-broken
state during node shutdown. In particular, their dependencies would be
already shut down, while the application itself was running, likely
with many errors.
The behaviour is to replicate what happens if the application finished
starting - with timeout set, we'll respect this timeout; otherwise
we'll wait forever.