Send the forker a stop message on port_close
This new message only has a minor effect in this patch: the forker
stops tracking the child process on port_close, so if the child
process is still running but stops later while the VM is alive, the
forker will no longer send an exit status message to the closed port.

The motivation for this patch is mostly just to set up the
communication mechanism, to attach more interesting behavior later
such as optionally killing the child process.