Unicode16 changed nfc and nfck functionality
Unicode16 added some codepoints that exposed flaws in the previous
implementation.

Qoute from spec
```
9.2 Normalization Contexts Requiring Care in Optimization
Starting with Unicode 16.0, there are several new characters (in the Kirat Rai,
Tulu-Tigalari, and Gurung Khema scripts) with normalization behavior
not seen in characters encoded in earlier versions of the Unicode
Standard. The normalization algorithm and the definitions of
normalization-related properties have not changed.

However, Unicode 16.0 is the first version which includes some composite characters
that can occur in NFC/NFKC strings, but when those characters occur in
a context directly following certain other characters, performing an
NFC or NFKC normalization will change those composite characters. (A
composite character has a Decomposition_Mapping (dm) value consisting
of a sequence of more than one character. In this case, the first
characters in their decompositions can combine with certain preceding
characters.) This situation is illustrated schematically in the
following table, using an arbitrary convention of square brackets to
indicate a composite character.
```