ort: Run ort scan on each push/pr and add cache
Instead of relying on our own custom scan-code implementation,
we rely on ort to do all analysis for us. The problem with using
ort is that a re-scan of all files takes more than 1h, which is
not acceptable, so this change introduces a way of scanning only
the files that have changed and rely on the cache to speed up
the scans.