[erts] Fix delayed receive trace
ERTS_PSFLG_ACTIVE_SYS should be set if any of ERTS_PSFLG_SIG_Q,
ERTS_PSFLG_NMSG_SIG_IN_Q, or ERTS_PSFLG_SYS_TASKS are set and ERTS_PSFLG_FREE
is not set. When a process is being scheduled out this might however not be
the case if current process:
* is being receive traced, and
* has called erts_proc_sig_fetch() which have brought in a message into
  the middle queue where no non-message signals exist (due to being receive
  traced), and
* is waiting in a 'receive' expression without clauses matching on messages
  (timed wait), or is just not entering 'receive' expressions for a very long
  time.
ERTS_PSFLG_SIG_Q will in this case be set, but not ERTS_PSFLG_ACTIVE_SYS,
so we need to set ERTS_PSFLG_ACTIVE_SYS. Otherwise, the message will not be
receive traced until another non-message signal arrives, or the process time
out and enter a 'receive' expression matching on messages.

If ERTS_PSFLG_NMSG_SIG_IN_Q and/or ERTS_PSFLG_SYS_TASKS have been set, it
has been done by another thread that later will set ERTS_PSFLG_ACTIVE_SYS and
wake us up, so we do not necessarily need to set ERTS_PSFLG_ACTIVE_SYS if
only ERTS_PSFLG_NMSG_SIG_IN_Q and/or ERTS_PSFLG_SYS_TASKS are set, but it
is good to do in order to schedule the process at once instead of waiting for
the wakeup.